Update example files and CMake configuration
- Replace the ONNX example with a new JPEG processing example that demonstrates a complete CPU anonymisation workflow.
- Update CMakeLists.txt to link the new JPEG example and remove the deprecated ONNX example.
- Enhance documentation in README to reflect the new example structure and clarify the purpose of each example.
- Add detailed comments in the new JPEG example to explain the workflow and processing steps.
